not working yet but getting closer. major redesign on the UI to allow it to be used for both edit and new, and only ahve 1 save point, to ensure no curruption of things, by trying to save timing windows before making an observation
